update conda install documents (#925)
Starting from v2.0.0.b4, `libdeepmd` and `lammps-dp` will decouple.
The idea is that both C++ API and LAMMPS are usually stable, so we do not need to build LAMMPS in every release. Also, CPU version and GPU version share the same API and LAMMPS itself does not need CUDA, so we do not need to build LAMMPS twice.